% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask_dir_create}
\alias{ask_dir_create}
\title{Ask: create directory if missing}
\usage{
ask_dir_create(
  dir = "/path/to/directory",
  title = "Directory check",
  preamble_exists_no = c("Directory {.field {dir}} does not exist yet.",
    "Do you want to create it?"),
  preamble_exists_yes = c("Directory {.field {dir}} already exists."),
  implications_yes_addendum = character(),
  implications_yes = c("Directory {.field {dir}} will be created"),
  implications_no_addendum = character(),
  implications_no = c("Nothing ;-)"),
  step = 0,
  steps_max = 0,
  reset = FALSE
)
}
\arguments{
\item{dir}{(\link{character} or \link{fs_path}) Directory in question.}

\item{title}{(\link{character}) Title for this "input block".}

\item{preamble_exists_no}{(\link{character}) Preamble text.}

\item{preamble_exists_yes}{(\link{character}) Preamble text.}

\item{implications_yes_addendum}{(\link{character}) Additional text for the
implications preamble text}

\item{implications_yes}{(\link{character}) Vector with implications}

\item{implications_no_addendum}{(\link{character}) Additional text for the
implications preamble text}

\item{implications_no}{(\link{character}) Vector with implications}

\item{step}{(\link{integer}) Denote which process step this is. See \code{\link[cli:h1]{cli::h1()}}.}

\item{steps_max}{(\link{integer}) Denote how many process steps there are in
total. See \code{\link[cli:h1]{cli::h1()}}.}

\item{reset}{[\link{logical}] Reset existing directories yes/no? Yes implies that
all subdirectories are deleted}
}
\value{

}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
path <- ask_dir_create(dir = fs::path(tempdir(), "a/b/c"))
path \%>\% fs::dir_exists()

ask_dir_create(dir = fs::path(tempdir(), "a/b/c"))
# Answer with an input of 2 ("No")

ask_dir_create(dir = fs::path(tempdir(), "a/b/c"))
# Answer with an input of 3 ("Let me start over")

ask_dir_create(dir = fs::path(tempdir(), "a/b/c"))
# Answer with an input of 4 ("Exit")
}
}
