% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clix-package.R
\docType{package}
\name{clix-package}
\alias{clix}
\alias{clix-package}
\title{clix: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
