% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CLix.scoring}
\alias{CLix.scoring}
\title{Corticolimbic Index (CLix) Function}
\usage{
CLix.scoring(
  dat = NA,
  reference.dat = NA,
  idname = "NPID",
  posHC.dat = NA,
  posHC.refdat = NA,
  crit.values = T
)
}
\arguments{
\item{dat}{The data provided by the user. The dataset must contain the column names: "H1", "H2", "C1", "C2", and "C3".    \cr
\cr H1 and H2 refer to the neurofibrillary tangle (NFT) counts in the hippocampal subsectors of CA1 and Subiculum; C1, C2 and C3 refer to the NFT counts in the cortical regions of Temporal, Parietal and Frontal.}

\item{reference.dat}{The reference dataset that can either be user provided or the built-in dataset FLAME-AD. \cr
\cr The FLAME-AD cohort represents the data from AD cases in the larger Florida Autopsied Multi-Ethnic series (n=1361).}

\item{idname}{The unique identifier for each case, the default is NPID.}

\item{posHC.dat}{The default is NA.}

\item{posHC.refdat}{The default is NA.}

\item{crit.values}{The default is TRUE. Used to toggle the calculation of the ratio of Hippocampal to Cortical means, R, for the reference dataset and print them out.}
}
\description{
This function will return a continuous measurement, corticolimbic index (CLix.score), along with a confidence interval. It will also assign an AD subtype (CLix.score.subtype) based on the Neurofibrillary Tangle (NFT) counts of five key brain regions for each case.
}
\keyword{CLix}
