% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{dir_reset}
\alias{dir_reset}
\title{Reset directory}
\usage{
dir_reset(dir)
}
\arguments{
\item{dir}{[\link{character}] Directory to reet}
}
\value{

}
\description{
Reset directory
}
\examples{
dir <- fs::path(tempdir(), "maindir") \%>\% fs::dir_create()
subdir <- fs::path(dir, "subdir") \%>\% fs::dir_create()
dir_reset(dir)
}
