% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
A data frame with 262 rows and 7 variables:
\describe{
\item{NPID}{Factor. Case identifier.}
\item{AD.subtype}{Factor with 3 levels (HpSp, Typical, Limbic) describing the Alzheimer's disease subtype.}
\item{CA1.NFT}{Numeric. The Neurofibrillary tangle (NFT) counts in the hippocampal subsectors of CA1.}
\item{Subiculum.NFT}{Numeric. The Neurofibrillary tangle (NFT) counts in the hippocampal subsectors of Subiculum.}
\item{Temporal.NFT}{Numeric. The Neurofibrillary tangle (NFT) counts in the cortical regions of Temporal.}
\item{Parietal.NFT}{Numeric. The Neurofibrillary tangle (NFT) counts in the cortical regions of Parietal.}
\item{Frontal.NFT}{Numeric. The Neurofibrillary tangle (NFT) counts in the cortical regions of Frontal.}
}
}
\usage{
example_data
}
\description{
An example data set containing AD information for 262 cases.
}
\keyword{datasets}
