% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{h1}
\alias{h1}
\title{Heading level 1 \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
h1(text, step = 0, steps_max = 0, .width = 80, .envir = parent.frame())
}
\arguments{
\item{text}{(\link{character}) Text of the heading. It can contain inline markup.}

\item{step}{(\link{integer}) Denote which process step this is.}

\item{steps_max}{(\link{integer}) Denote how many process steps there are in total}

\item{.width}{(\link{integer}) Line width. Currently not compatible with.
\code{\link[cli:cli_h1]{cli::cli_h1()}} but possibly usefule in the future.}

\item{.envir}{(\link{environment}) Environment to evaluate the glue expressions
in.}
}
\value{

}
\description{
Mainly a "decoupling wrapper" around \code{\link[cli:cli_h1]{cli::cli_h1()}} with some additional
features.
}
\examples{
h1("This is a title")
}
