% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{handle_input_}
\alias{handle_input_}
\title{Handle input (internal part) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
handle_input_(answer, type = c("inner", "outer"), envir = parent.frame())
}
\arguments{
\item{envir}{}
}
\value{

}
\description{
Handle input (internal part) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
