% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{handle_multi_lines}
\alias{handle_multi_lines}
\title{Handle multiple lines \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
handle_multi_lines(x, .envir = parent.frame())
}
\arguments{
\item{x}{(\link{character}) Vector of input that should be transformed to multi
lines in console output.}

\item{.envir}{(\link{environment}) Environment to evaluate the glue expressions
in.}
}
\value{

}
\description{
Handle multiple lines \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
