% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{HpSp.scoring}
\alias{HpSp.scoring}
\title{CLix Scoring Function}
\usage{
HpSp.scoring(
  dat = NA,
  reference.dat = NA,
  idname = "NPID",
  posHC.dat = NA,
  posHC.refdat = NA,
  crit.values = T
)
}
\arguments{
\item{dat}{the data containing cases to assign a score and AD subtype}

\item{reference.dat}{the reference data containing cases for which to score cases against, if NA nbM Flame cohort is used}

\item{idname}{the unique identifier for which to identify the cases, default is NPID}

\item{posHC.dat}{default is NA}

\item{posHC.refdat}{default is NA}

\item{crit.values}{default is TRUE}
}
\description{
This function will score and assign an AD subtype each case based on the distribution it's tangles
}
\keyword{hpsp}
