% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implications.R
\name{implications_input_yes}
\alias{implications_input_yes}
\title{Implications: input yes \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
implications_input_yes(
  addendum = character(),
  implications = character(),
  .envir = parent.frame()
)
}
\arguments{
\item{addendum}{(\link{character}) Addendum for preamble text}

\item{implications}{(\link{character}) Vector with implications. Will be turned
into a bullet list. See \code{\link[=ul]{ul()}}.}

\item{envir}{(\link{environment}) The environment in which to evaluate expression.}
}
\value{

}
\description{
Implications: input yes \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
implications_input_yes(
    addendum = "when running 'foo()'",
    implications = letters[1:3]
)
}
}
