% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_keep_reset_again_exit}
\alias{input_keep_reset_again_exit}
\title{User input: yes/no/again/exit \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
input_keep_reset_again_exit(
  title = character(),
  value = valid::valid_keep_reset_again_exit("reset"),
  force = FALSE
)
}
\arguments{
\item{title}{(\link{character}) Optional title before presenting the input
choices. Currently not "compatible as desired" between {\link{cli}} and
\link{select.list} and thus set to \code{character()} by default.}

\item{value}{[\link{character}] Value for non-interactive situations}

\item{force}{[\link{logical}] Force usage of value yes/no}
}
\value{

}
\description{
User input: yes/no/again/exit \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
input_keep_reset_again_exit()
input_keep_reset_again_exit(force = TRUE)
}
}
