% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbM_Flame_NFT.R
\docType{data}
\name{nbM_Flame_NFT}
\alias{nbM_Flame_NFT}
\title{nbM Flame Cohort}
\format{
A data frame with 1361 rows and 7 variables:
\describe{
\item{NPID}{factor, row number}
\item{AD.subtype}{factor with 3 levels describing Alzheimer's disease assigned subtype based on tangle burden/count}
\item{H1}{numeric}
\item{H2}{numeric}
\item{C1}{numeric}
\item{C2}{numeric}
\item{C3}{numeric}
}
}
\usage{
nbM_Flame_NFT
}
\description{
A data set containing AD information for 1361 cases in the nbM Flame Cohort
}
\keyword{datasets}
