% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw.R
\name{throw_error}
\alias{throw_error}
\title{Throw error \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
throw_error(
  msg = "Error code: 1234",
  .console = FALSE,
  .log = FALSE,
  .envir = parent.frame()
)
}
\arguments{
\item{msg}{(\link{character}) Error message. See \code{\link[stringr:str_glue]{stringr::str_glue()}}.}

\item{.console}{(\link{logical}) Output to console}

\item{.log}{(\link{logical}) Output to log (console or file depending on logging
settings)
TODO-20220201-1627: flesh out logging details}

\item{.envir}{(\link{environment}) Environment in which to evaluate msg expression}
}
\value{

}
\description{
Throw error \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
throw_error(msg = "This is an error")

error_code <- 1234
throw_error(msg = "This is an error (error code: {error_code})", )

throw_error(msg = "This is an error", .console = TRUE)
throw_error(msg = "This is an error", .log = TRUE)
}
}
