% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw.R
\name{throw_warning}
\alias{throw_warning}
\title{Throw warning \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
throw_warning(
  msg = "Warning code: 1234",
  .console = FALSE,
  .log = FALSE,
  .envir = parent.frame()
)
}
\arguments{
\item{msg}{(\link{character}) Warning message. See \code{\link[stringr:str_glue]{stringr::str_glue()}}.}

\item{.console}{(\link{logical}) Output to console}

\item{.log}{(\link{logical}) Output to log (console or file depending on logging
settings)
TODO-20220201-1627: flesh out logging details}

\item{.envir}{(\link{environment}) Environment in which to evaluate msg expression}
}
\value{

}
\description{
Throw warning \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
