% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{ul}
\alias{ul}
\title{Unordered list \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ul(items, .envir = parent.frame())
}
\arguments{
\item{items}{(\link{character}) If not NULL, then a character vector. Each element
of the vector will be one list item, and the list container will be closed
by default (see the .close argument).}

\item{.envir}{(\link{environment}) Environment to evaluate the glue expressions
in. It is also used to auto-close the container if .auto_close is TRUE.}
}
\value{

}
\description{
Pure "decoupling wrapper" around \code{\link[cli:cli_ul]{cli::cli_ul()}}
}
\examples{
ul(letters[1:3])
}
