% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{wrap_line}
\alias{wrap_line}
\title{Wrap lines \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
wrap_line(.x, .width = 80, .indent = 0, .envir = parent.frame(), .par = TRUE)
}
\arguments{
\item{.par}{}
}
\value{

}
\description{
Wrap lines \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
