% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{wrap_success}
\alias{wrap_success}
\title{Wrap success \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
wrap_success(.x, .width = 80, .envir = parent.frame())
}
\arguments{
\item{.envir}{}
}
\value{

}
\description{
Wrap success \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
