% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clix-package.R
\docType{package}
\name{clix-package}
\alias{clix}
\alias{clix-package}
\title{clix: A custom CLI implementation}
\description{
A custom CLI implementation with features for common CLI scenarios such as "ask user if a non-existing directory should be created or not". It tries to provide sensible bundles of a) messages (in the broader sense) b) querying for user input and c) executing the implications of a given user input
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
