% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLix.R
\docType{package}
\name{CLix}
\alias{CLix}
\title{Corticolimbic Index (CLix) Package}
\description{
The goal of CLix is to personalize the Alzheimer's disease (AD) subtype algorithm by expressing it as a
continuous numeric corticolimbic index (CLix). It was derived using tangle count distributions
from AD cases within the Florida Autopsy Multi-Ethnic series (FLAME-AD) to investigate clinical and demographic correlates.
}
\details{
The original algorithm was built upon a smaller cohort of AD cases (n=889) from a \href{https://doi.org/10.1016/S1474-4422(11)70156-9}{previous study}.
The algorithm was then improved and tested with the FLAME-AD\if{html}{\out{<sup>}}1,2\if{html}{\out{</sup>}} (Florida Autopsied Multi-Ethnic - Alzheimer's Disease)
cohort (n=1361). \cr
\cr
FLAME-AD was
derived from the State of Florida brain bank housed at the Mayo Clinic Florida. Donors
are derived through community engagement and participating Memory Disorder Clinics
in the State of Florida’s Alzheimer’s Disease Initiative who may register for autopsy
regardless of sex, race, or ethnicity.

\cr
\enumerate{
\item \href{https://link.springer.com/article/10.1007/s00401-018-1908-x}{Liesinger et. Al. Acta Neuropathologica 2018}
\cr
\item \href{https://alz-journals.onlinelibrary.wiley.com/doi/epdf/10.1016/j.jalz.2018.12.013}{Santos et. al. Alzheimer’s & Dementia 2019}
\cr
}

\cr Developed by the \href{https://www.mayo.edu/research/labs/translational-neuropathology/overview}{Translational Neuropathology Lab} led by Dr. Melissa E. Murray at Mayo Clinic Florida
}
