% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nbM_Flame_NFT}
\alias{nbM_Flame_NFT}
\title{FLAME-AD Cohort}
\format{
A data frame with 1361 rows and 7 variables:
\describe{
\item{NPID}{Factor. Case identifier.}
\item{AD.subtype}{Factor with 3 levels (HpSp, Typical, Limbic) describing the Alzheimer's disease subtype.}
\item{H1}{Numeric. The Neurofibrillary tangle (NFT) counts in the hippocampal subsectors of CA1.}
\item{H2}{Numeric. The Neurofibrillary tangle (NFT) counts in the hippocampal subsectors of Subiculum.}
\item{C1}{Numeric. The Neurofibrillary tangle (NFT) counts in the cortical regions of Temporal.}
\item{C2}{Numeric. The Neurofibrillary tangle (NFT) counts in the cortical regions of Parietal.}
\item{C3}{Numeric. The Neurofibrillary tangle (NFT) counts in the cortical regions of Frontal.}
}
}
\usage{
nbM_Flame_NFT
}
\description{
A dataset containing the AD information for the Florida Autopsy Multi-Ethnic series (FLAME-AD) Cohort (n=1361).
}
\keyword{datasets}
